Firebase react hooks is installed and working gloriously being tested on dashboar which is an excellent place to do it!
