+++ blitzkarte/package.json
@@ -17,7 +17,8 @@
     "gsap": "^3.10.3",
     "next": "^12.1.0",
     "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react-dom": "^17.0.2",
+    "react-firebase-hooks": "^5.0.3"
   },
   "devDependencies": {
     "@types/node": "^17.0.21",
